refactor:重构注册事件方式
