--- src/main/modules/close.ts
@@ -1,8 +1,0 @@
-import { ipcMain, app } from "electron";
-
-// 退出
-export function quitWindow() {
-  ipcMain.on("quit:application", () => {
-    app.quit()
-  });
-}
