--- src/main/modules/switch.ts
@@ -1,8 +1,0 @@
-import { BrowserWindow } from 'electron';
-import { ElectronWindowType } from '../window-type';
-import WindowFactory from '../window';
-
-export function switchWindow(winType: ElectronWindowType, win: BrowserWindow) {
-  win.close()
-  return WindowFactory.createWindow(winType)
-}
