+++ src/main/window/windowFactory.ts
